Add files via upload
In this project, you’ll plan and build an Autofill Search Bar for various types of fruit!You may be familiar with autofill from a search engine like Google. When you type anything into Google, a drop down appears with suggestions for what you may want to search for. If you click on one of these suggestions, it populates the search bar with that suggestion.

Building search bars like this is a common task for front-end engineers. You’ll be building a small scale search bar which suggests various fruits based on user input. We’ll provide a list of fruits and a mock up of how the page should look. You’ll do the rest. If a user’s typed input is contained anywhere in one of the fruits in the list, then a drop down will appear containing all the fruits which match. As a user continues to add letters to their input, new dropdowns will populate. When a user clicks on one of these options, it will populate in the search bar.